Multiple pp_check plots per species

Multiple pp_check plots per species

Usage

multi_pp_check(
  object,
  plotfun = "dens_overlay",
  species = NULL,
  ndraws = NULL,
  grid_args = list(...),
  ...
)

Arguments

object

The jsdmStanFit model object

plotfun

The ppc plot function to use, given as a character string. The default is to call ppc_dens_overlay. Can be specified as either the entire name of function as a character string or without the ppc_ prefix.

species

Which species should be included, by default all

ndraws

How many draws should be used within the plots

grid_args

Optional list of arguments that are passed to gridExtra::arrangeGrob() (nrow, ncol, widths, etc.)

...

Other options passed to pp_check

Value

An object of class "bayesplot_grid", for more information see bayesplot::bayesplot_grid()